UI update

+++ include/MyBot.h
@@ -8,8 +8,8 @@ class MyBot : public Chess::AbstractBot
 {
 	public:
 		bool GetTurn(Turn &t,Chessboard &board);
-        void GetTreeMark(double &TreeMark , bool &TreeResult , Chessboard &board , bool MinMax , unsigned TTL );
-        void GetPositionMark(double &PositionMark , Chessboard &board);
+        void GetTreeMark(float &TreeMark , bool &TreeResult , Chessboard &board , bool MinMax , unsigned TTL );
+        void GetPositionMark(float &PositionMark , Chessboard &board);
 };
